Remove qualifiers, simplify logic for unambiguous references
The qualifiers and related functions are not necessary due to UI
support for moving references. Checking ambiguous references required
many steps that have been simplified. Control over what is
displayed is mostly delegated to the replacer logic with
a flag in the DataflowRef to indicate if the ref is ambiguous.